Add diagnostic prints
So here's what's important:

getFromIndex() returns the index of the first character in the
currently-modified word if typing consecutive non-whitespace characters.
If the cursor moves to a different point in the string, getFromIndex()
returns the index of that point.  When a space is inserted,
getFromIndex() always returns the index of that space.